2024-2025

+++ Final assignment/train.py
@@ -126,7 +126,7 @@ def main(args):
         for i, (images, labels) in enumerate(train_dataloader):
             images, labels = images.to(device), labels.to(device)
 
-            labels = labels.squeeze(1)  # Remove channel dimension
+            labels = labels.long().squeeze(1)  # Remove channel dimension
 
             optimizer.zero_grad()
             outputs = model(images)
@@ -147,7 +147,7 @@ def main(args):
             for i, (images, labels) in enumerate(valid_dataloader):
                 images, labels = images.to(device), labels.to(device)
 
-                labels = labels.squeeze(1)  # Remove channel dimension
+                labels = labels.long().squeeze(1)  # Remove channel dimension
 
                 outputs = model(images)
                 loss = criterion(outputs, labels)
